bot_compiler & extraction samples
Within the folder you will find a .py that compiles 4 different bots, each for a different webpage:
- workwise (with selenium)
- indeed (with selenium)
- stellenanzeigen (with requests)
- jobteaser (with requests)
Also, you will find some extraction samples. The prefix is a correlation of the date it was extracted (i.e. YYYYMMDD). So the earlier ones (from the 3rd of April) are the most complete in terms of columns extracted, as well as a similar structure for each webpage extraction, so that they all can be potentially joined in just one excel.